backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
psycopg[binary]>=3.2.1
psycopg2-binary>=2.9.9
pyjwt==2.9.0
alembic==1.13.2
python-dotenv==1.0.1
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
openai>=1.10.0
email-validator>=2.0.0
aiokafka>=0.10.0
